+++ pom.xml
@@ -2,7 +2,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.telefonica.euro_iaas</groupId>
+    <groupId>com.telefonica.fiware</groupId>
     <artifactId>commons</artifactId>
     <packaging>jar</packaging>
     <name>Commons</name>
