+++ src/main/java/com/telefonica/fiware/commons/dao/AbstractSearchCriteria.java
@@ -22,7 +22,7 @@
  * </p>
  */
 
-package com.telefonica.euro_iaas.commons.dao;
+package com.telefonica.fiware.commons.dao;
 
 /**
  * Provides a set of common fields in search criteria.
