+++ src/main/java/com/telefonica/fiware/commons/dao/AlreadyExistsEntityException.java
@@ -22,7 +22,7 @@
  * </p>
  */
 
-package com.telefonica.euro_iaas.commons.dao;
+package com.telefonica.fiware.commons.dao;
 
 /**
  * This exception will be launched when a duplicated entity is persisted.
